+++ Assets/Negi0109/ColorGradientToTexture/Scripts/ColorFilters/Add.cs
@@ -6,14 +6,14 @@ namespace Negi0109.ColorGradientToTexture.Filters
     [System.Serializable]
     public class Add : ColorFilter
     {
-        public float value;
-        public override float Evaluate(float v) => v + value;
+        public double value;
+        public override double Evaluate(double v) => v + value;
 
         public override bool Editor()
         {
             return Watch(
                 () => value,
-                () => { value = EditorGUILayout.FloatField("value", value); }
+                () => { value = EditorGUILayout.DoubleField("value", value); }
             );
         }
     }
