+++ Assets/Negi0109/ColorGradientToTexture/Scripts/ColorFilters/OneMinus.cs
@@ -3,7 +3,7 @@ namespace Negi0109.ColorGradientToTexture.Filters
     [System.Serializable]
     public class OneMinus : ColorFilter
     {
-        public override float Evaluate(float v) => 1 - v;
+        public override double Evaluate(double v) => 1 - v;
 
         public override bool Editor() => false;
     }
