+++ Assets/Negi0109/ColorGradientToTexture/Scripts/ColorFilters/Threshold.cs
@@ -5,14 +5,14 @@ namespace Negi0109.ColorGradientToTexture.Filters
     [System.Serializable]
     public class Threshold : ColorFilter
     {
-        public float value;
-        public override float Evaluate(float v) => v >= value ? 1 : 0;
+        public double value;
+        public override double Evaluate(double v) => v >= value ? 1 : 0;
 
         public override bool Editor()
         {
             return Watch(
                 () => value,
-                () => { value = EditorGUILayout.Slider("value", value, 0f, 1f); }
+                () => { value = EditorGUILayout.Slider("value", (float)value, 0f, 1f); }
             );
         }
     }
