+++ Assets/Negi0109/ColorGradientToTexture/Tests/ColorFilters/AddTest.cs
@@ -4,9 +4,9 @@ namespace Negi0109.ColorGradientToTexture.Tests.ColorFilterTests
 {
     public class AddTest
     {
-        [TestCase(1f, 0.5f, 1.5f)]
-        [TestCase(1f, 1f, 2f)]
-        public void Evaluate(float value, float v, float expected)
+        [TestCase(1, 0.5, 1.5)]
+        [TestCase(1, 1, 2)]
+        public void Evaluate(double value, double v, double expected)
         {
             var filter = new Filters.Add() { value = value };
