+++ Assets/Negi0109/ColorGradientToTexture/Tests/ColorFilters/ScaleTest.cs
@@ -4,9 +4,9 @@ namespace Negi0109.ColorGradientToTexture.Tests.ColorFilterTests
 {
     public class ScaleTest
     {
-        [TestCase(2, 0.5f, 1f)]
-        [TestCase(2, 1f, 2f)]
-        public void Evaluate(float scale, float v, float expected)
+        [TestCase(2, 0.5, 1)]
+        [TestCase(2, 1, 2)]
+        public void Evaluate(double scale, double v, double expected)
         {
             var filter = new Filters.Scale() { value = scale };
